src/requirements.txt
fastapi==0.115.5
uvicorn==0.32.0
markdown==3.7
xhtml2pdf==0.2.16
python-docx==1.1.2
redis==5.2.0